backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
qdrant-client
openai
python-dotenv
psycopg2-binary
requests
beautifulsoup4
markdown
